+++ src/app/context/FirebaseContext/hooks/useFirebaseStorageHelper.tsx
@@ -1,8 +1,10 @@
 import { ref, uploadBytes, getDownloadURL, deleteObject, listAll, FirebaseStorage } from "firebase/storage";
-
-
+import { compressImage, getBase64FileSizeInKB } from "@/app/utilities/PicturesHelper/PictureHelper";
+import { FiletoURL, urlToFile } from "@/app/utilities/FilesHelper/FilesHelper";
+import { getCurrentDateTime } from "@/app/utilities/DatesHelper/Dateshelper";
 export interface FirebaseStorageHelper {
-    uploadFile: (file: any, filePath: string) => Promise<string>;
+    uploadImage: (file: File, filePath: string, qualitycompressed?: number | undefined) => Promise<string>
+    uploadFile: (file: File, filePath: string) => Promise<string>;
     updateFile: (file: File, filePath: string) => Promise<string>;
     listFilesAndUrls: (path: string) => Promise<any>;
     downloadFile: (filePath: string) => Promise<string>;
@@ -10,12 +12,27 @@ export interface FirebaseStorageHelper {
     storage: FirebaseStorage | null;
 }
 
-const useFirebaseStorageHelper = (storage: FirebaseStorage | null):FirebaseStorageHelper => {
+const useFirebaseStorageHelper = (storage: FirebaseStorage | null): FirebaseStorageHelper => {
 
-    const uploadFile = async (file: any, filePath: string) => {
+    const uploadImage = async (file: File, filePath: string, qualitycompressed?: number) => {
+        const fileUrl = await FiletoURL(file);
+        const originalSize = await getBase64FileSizeInKB(fileUrl);
+        console.log(`Tamaño original: ${originalSize} KB`);
+        const imageCompressed = await compressImage(fileUrl, qualitycompressed);
+        if (imageCompressed) {
+            const sizeInKB = await getBase64FileSizeInKB(imageCompressed);
+            console.log(`Tamaño de la imagen: ${sizeInKB} KB`);
+            const filecompressedImage = await urlToFile(imageCompressed, file.name, file.type);
+            return await uploadFile(filecompressedImage, filePath);
+        } else {
+            return uploadFile(file, filePath);
+        }
+    };
+
+    const uploadFile = async (file: File, filePath: string) => {
         try {
             if (storage) {
-                const storageRef = ref(storage, filePath);
+                const storageRef = ref(storage, filePath+getCurrentDateTime());
                 const snapshot = await uploadBytes(storageRef, file);
                 return await getDownloadURL(snapshot.ref);
             }
@@ -90,7 +107,8 @@ const useFirebaseStorageHelper = (storage: FirebaseStorage | null):FirebaseStora
         updateFile,
         listFilesAndUrls,
         downloadFile,
-        deleteFile
+        deleteFile,
+        uploadImage
     }
 }
 export default useFirebaseStorageHelper