+++ src/app/main-page/accounting/personalInvoices/invoices/components/types.ts
@@ -1,5 +1,9 @@
+import { HistoryRow } from "@/app/mappings/billinghistory/billinghistory.types"
 
+  
 export interface InvoicesFormProps {
-  type: 'create' | 'update',
+  dataEdit?: HistoryRow | null
   layoutMatrix: number[][] | undefined
+  externalSubmitRef?: React.RefObject<(() => void | Promise<void>) | null>
 }
+